+++ platforms/gradle/foundry-gradle-plugin/src/main/kotlin/foundry/gradle/stats/ModuleStats.kt
@@ -223,7 +223,7 @@ public object ModuleStatsTasks {
 
             // TODO do we need to check the gradle properties too?
             // TODO move to the task action once we have these in ModuleTopography
-            val androidResources = extension.buildFeatures.androidResources == true
+            val androidResources = extension.androidResources.enable
             val viewBinding = extension.buildFeatures.viewBinding == true
             if (viewBinding) {
               addGeneratedSources()
@@ -448,15 +448,11 @@ internal object StatsUtils {
   fun parseProjectDeps(text: String): Set<String> {
     val deps = mutableSetOf<String>()
     text.lineSequence().forEach { line ->
-      if ("(projects." in line) {
+      if ("project(\"" in line) {
         // testFixtures*( are just for gradle module metadata and not actually a dependency
         if ("testFixturesApi(" in line) return@forEach
         if ("testFixturesImplementation(" in line) return@forEach
-        deps +=
-          line
-            .substringAfter("(projects.")
-            .substringBefore(")")
-            .substringBefore(".dependencyProject")
+        deps += line.substringAfter("project(\"").substringBefore("\")")
       }
     }
     return deps.toSortedSet()